Change `seasons` property of `Team` from `MutableMap` to `MutableList`.
Originally, idea was that the seasons would be searchable by mapping from year to season. Although this may be an intuitive data structure for relating a year to a specific season, searching seasons for a specific year is a viable algorithm for this application. Leveraging the existing metadata of a season (originally intended for redundant, doubly-linked referencing) also allows future databases to avoid storing an awkward key-index pair.